# Optimized very long case insensitive alternations