--- a/model/labels/regexp.go
+++ b/model/labels/regexp.go
@@ -553,14 +553,52 @@ func (m *anyStringMatcher) Matches(s string) bool {
 // this specific case, when we have many strings to match against we can use a map instead
 // of iterating over the list of strings.
 func optimizeEqualStringMatchers(input StringMatcher, threshold int) StringMatcher {
-	values, caseSensitive := findEqualStringMatchers(input)
-	if values == nil {
+	var (
+		caseSensitive    bool
+		caseSensitiveSet bool
+		numValues        int
+	)
+
+	// Analyse the input StringMatcher to count the number of occurrences
+	// and ensure all of them have the same case sensitivity.
+	analyseCallback := func(matcher *equalStringMatcher) bool {
+		// Ensure we don't have mixed case sensitivity.
+		if caseSensitiveSet && caseSensitive != matcher.caseSensitive {
+			return false
+		} else if !caseSensitiveSet {
+			caseSensitive = matcher.caseSensitive
+			caseSensitiveSet = true
+		}
+
+		numValues++
+		return true
+	}
+
+	if !findEqualStringMatchers(input, analyseCallback) {
 		return input
 	}
 
-	if len(values) < threshold {
+	// If the number of values found is less than the threshold, then we should skip the optimization.
+	if numValues < threshold {
 		return input
 	}
+
+	// Parse again the input StringMatcher to extract all values and storing them.
+	// We can skip the case sensitivity check because we've already checked it and
+	// if the code reach this point then it means all matchers have the same case sensitivity.
+	values := make(map[string]struct{}, numValues)
+
+	// Ignore the return value because we already iterated over the input StringMatcher
+	// and it was all good.
+	findEqualStringMatchers(input, func(matcher *equalStringMatcher) bool {
+		if caseSensitive {
+			values[matcher.s] = struct{}{}
+		} else {
+			values[strings.ToLower(matcher.s)] = struct{}{}
+		}
+
+		return true
+	})
 
 	return &equalMultiStringMatcher{
 		values:        values,
@@ -568,58 +606,33 @@ func optimizeEqualStringMatchers(input StringMatcher, threshold int) StringMatch
 	}
 }
 
-// findEqualStringMatchers analyze the input StringMatcher and return a map with all
-// equalStringMatcher values if and only if the input StringMatcher is composed by an
-// alternation of equalStringMatcher with the same case sensitivity.
-func findEqualStringMatchers(input StringMatcher) (values map[string]struct{}, caseSensitive bool) {
+// findEqualStringMatchers analyze the input StringMatcher and calls the callback for each
+// equalStringMatcher found. Returns true if and only if the input StringMatcher is *only*
+// composed by an alternation of equalStringMatcher.
+func findEqualStringMatchers(input StringMatcher, callback func(matcher *equalStringMatcher) bool) bool {
 	orInput, ok := input.(orStringMatcher)
 	if !ok {
-		return nil, false
+		return false
 	}
-
-	values = map[string]struct{}{}
 
 	for _, m := range orInput {
 		switch casted := m.(type) {
 		case orStringMatcher:
-			subValues, subCaseSensitive := findEqualStringMatchers(m)
-			if subValues == nil {
-				return nil, false
-			}
-
-			// Ensure we don't have mixed case sensitivity.
-			if len(values) > 0 && caseSensitive != subCaseSensitive {
-				return nil, false
-			} else if len(values) == 0 {
-				caseSensitive = subCaseSensitive
-			}
-
-			// Merge the values with our owns.
-			for v := range subValues {
-				values[v] = struct{}{}
+			if !findEqualStringMatchers(m, callback) {
+				return false
 			}
 
 		case *equalStringMatcher:
-			// Ensure we don't have mixed case sensitivity.
-			if len(values) > 0 && caseSensitive != casted.caseSensitive {
-				return nil, false
-			} else if len(values) == 0 {
-				caseSensitive = casted.caseSensitive
-			}
-
-			// Store the value (lower case if matcher is case insensitive).
-			if caseSensitive {
-				values[casted.s] = struct{}{}
-			} else {
-				values[strings.ToLower(casted.s)] = struct{}{}
+			if !callback(casted) {
+				return false
 			}
 
 		default:
 			// It's not an equal string matcher, so we have to stop searching
 			// cause this optimization can't be applied.
-			return nil, false
+			return false
 		}
 	}
 
-	return
+	return true
 }