When a regex matcher contains a long list of case insensitive alternations, the `FastRegexMatcher` optimise the matching using the `StringMatcher` returned by `stringMatcherFromRegexp()`. In particular, we will be in the case there are 1 or more nested `orStringMatcher` each one with multiple `equalStringMatcher`. All the `equalStringMatcher` are evaluated sequentially, so it's like iterating over a list of string values and compare each of them against the input.

In this specific case, when there are many alternations, it may be faster to put all values in a map and then lookup the values in the map to find if there's anyone of them matching. This is what this PR does.

To do it, I've introduced the function `optimizeEqualStringMatchers()` which optimise the output of `stringMatcherFromRegexp()` exactly for the case described above. The reason why it's difficult to optimise directly `stringMatcherFromRegexp()` is because an alternate regex could be unrolled as a set of alternate + concat, so we have to first unroll it and then check if we can apply the optimization.

I've introduced a couple of new cases to the benchmark. This optimization is valuable when the regex contains a long list of case insensitive alternations (which is the specific case I'm trying to optimise):

```
name                                                  old time/op    new time/op    delta
FastRegexMatcher/foo-12                                 7.05ns ± 2%    7.29ns ± 1%     ~     (p=0.063 n=3+3)
FastRegexMatcher/^foo-12                                7.82ns ± 7%    7.19ns ± 0%     ~     (p=0.166 n=3+3)
FastRegexMatcher/(foo|bar)-12                           9.15ns ±11%    9.58ns ± 4%     ~     (p=0.484 n=3+3)
FastRegexMatcher/foo.*-12                               32.1ns ±10%    29.2ns ± 3%     ~     (p=0.252 n=3+3)
FastRegexMatcher/.*foo-12                               32.3ns ± 1%    30.7ns ± 2%   -4.85%  (p=0.013 n=3+3)
FastRegexMatcher/^.*foo$-12                             35.8ns ±10%    31.1ns ± 2%     ~     (p=0.110 n=3+3)
FastRegexMatcher/^.+foo$-12                             32.8ns ± 2%    31.1ns ± 2%   -5.26%  (p=0.012 n=3+3)
FastRegexMatcher/.*-12                                  28.8ns ± 7%    27.1ns ± 2%     ~     (p=0.218 n=3+3)
FastRegexMatcher/.+-12                                  31.3ns ± 5%    27.9ns ± 1%  -10.90%  (p=0.048 n=3+3)
FastRegexMatcher/foo.+-12                               29.9ns ± 1%    29.2ns ± 2%     ~     (p=0.100 n=3+3)
FastRegexMatcher/.+foo-12                               33.8ns ± 5%    31.8ns ± 4%     ~     (p=0.134 n=3+3)
FastRegexMatcher/.*foo.*-12                             86.4ns ± 5%    81.2ns ± 2%     ~     (p=0.144 n=3+3)
FastRegexMatcher/(?i:foo)-12                            22.9ns ± 9%    22.8ns ± 2%     ~     (p=0.929 n=3+3)
FastRegexMatcher/(?i:(foo|bar))-12                      47.1ns ± 6%    43.6ns ± 3%     ~     (p=0.122 n=3+3)
FastRegexMatcher/(?i:(foo1|foo2|bar))-12                66.8ns ± 1%    67.1ns ± 3%     ~     (p=0.839 n=3+3)
FastRegexMatcher/(?i:(foo1|foo2|aaa|bbb|ccc|ddd|e-12     439ns ±10%     229ns ± 7%  -47.90%  (p=0.007 n=3+3)
FastRegexMatcher/(prometheus|api_prom)_api_v1_.+-12     46.7ns ± 3%    43.8ns ± 3%     ~     (p=0.051 n=3+3)
FastRegexMatcher/((fo(bar))|.+foo)-12                   66.4ns ±13%    61.2ns ± 2%     ~     (p=0.347 n=3+3)
FastRegexMatcher/zQPbMkNO|NNSPdvMi|iWuuSoAl|qbvKM-12     159ns ± 9%     168ns ± 3%     ~     (p=0.315 n=3+3)
FastRegexMatcher/(?i:(zQPbMkNO|NNSPdvMi|iWuuSoAl|-12    1.53µs ± 1%    0.23µs ± 4%  -85.23%  (p=0.000 n=3+3)

name                                                  old alloc/op   new alloc/op   delta
FastRegexMatcher/foo-12                                  0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/^foo-12                                 0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/(foo|bar)-12                            0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/foo.*-12                                0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/.*foo-12                                0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/^.*foo$-12                              0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/^.+foo$-12                              0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/.*-12                                   0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/.+-12                                   0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/foo.+-12                                0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/.+foo-12                                0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/.*foo.*-12                              0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/(?i:foo)-12                             0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/(?i:(foo|bar))-12                       0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/(?i:(foo1|foo2|bar))-12                 0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/(?i:(foo1|foo2|aaa|bbb|ccc|ddd|e-12     0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/(prometheus|api_prom)_api_v1_.+-12      0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/((fo(bar))|.+foo)-12                    0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/zQPbMkNO|NNSPdvMi|iWuuSoAl|qbvKM-12     0.00B          0.00B          ~     (zero variance)
FastRegexMatcher/(?i:(zQPbMkNO|NNSPdvMi|iWuuSoAl|-12     0.00B          0.00B          ~     (zero variance)

name                                                  old allocs/op  new allocs/op  delta
FastRegexMatcher/foo-12                                   0.00           0.00          ~     (zero variance)
FastRegexMatcher/^foo-12                                  0.00           0.00          ~     (zero variance)
FastRegexMatcher/(foo|bar)-12                             0.00           0.00          ~     (zero variance)
FastRegexMatcher/foo.*-12                                 0.00           0.00          ~     (zero variance)
FastRegexMatcher/.*foo-12                                 0.00           0.00          ~     (zero variance)
FastRegexMatcher/^.*foo$-12                               0.00           0.00          ~     (zero variance)
FastRegexMatcher/^.+foo$-12                               0.00           0.00          ~     (zero variance)
FastRegexMatcher/.*-12                                    0.00           0.00          ~     (zero variance)
FastRegexMatcher/.+-12                                    0.00           0.00          ~     (zero variance)
FastRegexMatcher/foo.+-12                                 0.00           0.00          ~     (zero variance)
FastRegexMatcher/.+foo-12                                 0.00           0.00          ~     (zero variance)
FastRegexMatcher/.*foo.*-12                               0.00           0.00          ~     (zero variance)
FastRegexMatcher/(?i:foo)-12                              0.00           0.00          ~     (zero variance)
FastRegexMatcher/(?i:(foo|bar))-12                        0.00           0.00          ~     (zero variance)
FastRegexMatcher/(?i:(foo1|foo2|bar))-12                  0.00           0.00          ~     (zero variance)
FastRegexMatcher/(?i:(foo1|foo2|aaa|bbb|ccc|ddd|e-12      0.00           0.00          ~     (zero variance)
FastRegexMatcher/(prometheus|api_prom)_api_v1_.+-12       0.00           0.00          ~     (zero variance)
FastRegexMatcher/((fo(bar))|.+foo)-12                     0.00           0.00          ~     (zero variance)
FastRegexMatcher/zQPbMkNO|NNSPdvMi|iWuuSoAl|qbvKM-12      0.00           0.00          ~     (zero variance)
FastRegexMatcher/(?i:(zQPbMkNO|NNSPdvMi|iWuuSoAl|-12      0.00           0.00          ~     (zero variance)
```


I've also re-run the fuzz tests for few minutes (they don't run in CI) and they reported no failures.