@pracucci 2023-03-02
> Also as a future improvement, I would consider replacing the

👍 But in a separate PR :)